// Backend Projects/hiring-management-backend-development/src/feedback/enum/feedback.enum.ts
export enum FeedbackFormType {
    technicalAutomotive = "Technical Automotive",
    technicalRnD = "Technical R&D",
    managerial = "Managerial",
    relationship = "Relationship",
    hr = "HR"
}

export enum MaritalStatus {
    married = "Married",
    unmarried = "Unmarried",
}

export enum AutomotiveBackground {
    yes = "Yes",
    no = "No",
}