// Backend Projects/hiring-management-backend-development/src/candidates/enum/candidate.enum.ts
export enum CandidateFlag {
    duplicate = "Duplicate",
    blacklisted = "Blacklisted"
}

export enum CandidateStatus {
    new = "New",
    shortlisted = "Shortlisted",
    notShortlisted = "Not Shortlisted",
    scheduled = "Scheduled",
    selected = "Selected", 
    notSelected = "Not Selected",
    hold = "Hold",
    inprogress = "In Progress",
    skipped = "Skipped"
}

export enum CandidateFilter {
    new = "New",
    shortlisted = "Shortlisted",
    notShortlisted = "Not Shortlisted",
    scheduled = "Scheduled",
    selected = "Selected", 
    notSelected = "Not Selected",
    hold = "Hold",
    inprogress = "In Progress",
    ongoing = "Ongoing",
    rejected = "Rejected"
}